support for roles in addition to already implemented authorities

+++ src/main/java/com/derbyware/registration/repositories/AuthorityRepository.java
@@ -4,9 +4,7 @@ import com.derbyware.registration.entities.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority, BigInteger> {
+public interface AuthorityRepository extends JpaRepository<Authority, Long> {
 	public Authority findAuthorityByName(String name);
 }
